% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{Query}
\alias{Query}
\title{Create \code{Query} object.}
\usage{
Query(
  datasetId = NULL,
  assemblyId = NULL,
  snpId = NULL,
  referenceName = NULL,
  start = NULL,
  end = NULL,
  geneSymbol = NULL
)
}
\arguments{
\item{datasetId}{Dataset ID - \code{bipmed-wes-phase2}}

\item{assemblyId}{Reference genome version - \code{hg19}, \code{GRCh38}}

\item{snpId}{NCBI dbSNP database ID - \code{rs35735053}}

\item{referenceName}{Chromosome - \code{1}, \code{X}}

\item{start}{0-based start position (inclusive) - \code{65000}}

\item{end}{end position (inclusive) - \code{70000}}

\item{geneSymbol}{gene name (case-insensitive) - \code{SCN1A}, \code{brca1}}
}
\value{
\code{Query} object.
}
\description{
Create \code{Query} object.
}
\examples{
# gene symbol
Query(geneSymbol = "SCN1A")
# genomic range
Query(referenceName = 1, start = 65000, end = 70000)
# genomic position
Query(referenceName = 1, start = 7737651)
# dbSNP ID
Query(snpId = "rs35735053")
}
