% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search}
\alias{search}
\alias{search.Query}
\title{Search for variants given a query.}
\usage{
search(x, ...)

\method{search}{Query}(x, start = 1, length = -1,
  URL = "https://bcbcloud.fcm.unicamp.br/bipmed/datatables")
}
\arguments{
\item{x}{\code{Query} object}

\item{start}{1-based page begin}

\item{length}{page length - \code{-1} for all variants (default)}

\item{URL}{server API endpoint - default to BraVE server}
}
\value{
a list of \code{Variant} objects.
}
\description{
Search for variants given a query.
}
\examples{
search(Query(geneSymbol = "SCN1A"), start = 1, length = 10)
search(Query(geneSymbol = "SCN1A"), start = 11, length = 10)
search(Query(referenceName = 1, start = 65000, end = 70000))
search(Query(referenceName = 1, start = 7737651))
search(Query(snpId = "rs35735053"))
}
