% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_brave.R
\name{search_brave}
\alias{search_brave}
\title{Search BraVE server.}
\usage{
search_brave(queries, start = 1, length = -1,
  URL = "https://bcbcloud.fcm.unicamp.br/brave/search",
  simplify = TRUE)
}
\arguments{
\item{queries}{gene symbols, genomic ranges, genomic position, dbSNP ID, mixed}

\item{start}{1-based page begin}

\item{length}{page length - \code{-1} for all variants (default)}

\item{URL}{server API endpoint - default to BraVE server}

\item{simplify}{return \code{data.frame} (default), otherwise list of \code{Variant} objects.}
}
\value{
variants which match the search queries
}
\description{
Search BraVE server.
}
\examples{
search_brave("SCN1A", start = 1, length = 10)
search_brave("scn1a", start = 11, length = 10)
search_brave("1:65000-70000")
search_brave("1:7737651")
search_brave("rs35735053")
search_brave(c("SCN1A", "1:65000-70000", "1:7737651", "rs35735053"))
}
