% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant.R
\name{Variant}
\alias{Variant}
\title{Create \code{Variant} object. Internal use.}
\usage{
Variant(snpIds, datasetId, assemblyId, totalSamples, referenceName, start,
  referenceBases, alternateBases, geneSymbol, alleleFrequency, sampleCount,
  coverage, genotypeQuality, clnsig, hgvs, type)
}
\arguments{
\item{snpIds}{Variant IDs}

\item{datasetId}{Dataset ID}

\item{assemblyId}{Reference genome version}

\item{totalSamples}{Total samples in dataset}

\item{referenceName}{Chromosome (VCF \code{CHROM} column)}

\item{start}{0-based start position (VCF \code{POS} column)}

\item{referenceBases}{Reference bases (VCF \code{REF} column)}

\item{alternateBases}{Alternate bases (VCF \code{ALT} column)}

\item{geneSymbol}{Gene symbol (Annotated VCF \code{ANN} column)}

\item{alleleFrequency}{Allele frequency (VCF \code{AF INFO} field)}

\item{sampleCount}{Samples that have variant (VCF \code{AF INFO} field)}

\item{coverage}{Distribution of coverage (VCF \code{DP} field)}

\item{genotypeQuality}{Distribution of genotype quality (VCF \code{GQ} field)}

\item{clnsig}{Clinical significance (Annotated VCF \code{CLNSIG} column)}

\item{hgvs}{Variant using HGVS notation - DNA level (Annotated VCF \code{ANN} field)}

\item{type}{Feature type - e.g. transcript, motif, miRNA, etc. (Annotated VCF \code{ANN} field)}
}
\value{
a \code{Variant} object.
}
\description{
Create \code{Variant} object. Internal use.
}
